refactor: update import paths for Exercise-related types and interfaces
