+++ app/src/lib/exercise_data/calisthenics.ts
@@ -1,5 +1,5 @@
+import type { ExerciseDetails } from "$lib/exercises";
 import { Muscles } from "$lib/muscles";
-import type { ExerciseDetails } from "$lib/types";
 
 /**
  * A collection of calisthenics exercises that can be used to generate workouts.
