+++ app/src/lib/exercise_data/machines.ts
@@ -1,6 +1,6 @@
 import { Equipment } from "$lib/equipment";
+import type { ExerciseDetails } from "$lib/exercises";
 import { Muscles } from "$lib/muscles";
-import type { ExerciseDetails } from "$lib/types";
 
 /**
  * A collection of predefined exercises that can be used to generate workouts.
